% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{hist.bewohner.by.stadtteil}
\alias{hist.bewohner.by.stadtteil}
\title{WG-Größe pro qm nach Stadtteil}
\usage{
\method{hist}{bewohner.by.stadtteil}(Daten, ordered = F)
}
\arguments{
\item{Daten}{}

\item{ordered}{}
}
\description{
WG-Größe pro qm nach Stadtteil
}
