% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{hist.count.by.stadtteil}
\alias{hist.count.by.stadtteil}
\title{Anzahl Wohnungen nach Stadtteil}
\usage{
\method{hist}{count.by.stadtteil}(Daten, ordered = F)
}
\arguments{
\item{Daten}{}

\item{ordered}{}
}
\description{
Anzahl Wohnungen nach Stadtteil
}
