% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{hist.geschl.by.stadtteil}
\alias{hist.geschl.by.stadtteil}
\title{Geschlechterverhältnisse nach Stadtteil}
\usage{
\method{hist}{geschl.by.stadtteil}(Daten, ordered = F)
}
\arguments{
\item{Daten}{}

\item{ordered}{}
}
\description{
Geschlechterverhältnisse nach Stadtteil
}
