% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{hist.mieteproqm.by.stadtteil}
\alias{hist.mieteproqm.by.stadtteil}
\title{Preis pro qm nach Stadtteil}
\usage{
\method{hist}{mieteproqm.by.stadtteil}(Daten, ordered = F)
}
\arguments{
\item{Daten}{}

\item{ordered}{}
}
\description{
Preis pro qm nach Stadtteil
}
