% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pie.bewohner.ges}
\alias{pie.bewohner.ges}
\title{Piechart zur gesuchten Geschlechtsverteilung}
\usage{
pie.bewohner.ges(Daten)
}
\arguments{
\item{Daten}{}
}
\description{
Piechart zur gesuchten Geschlechtsverteilung
}
