% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pie.bewohner}
\alias{pie.bewohner}
\title{Piechart zu Geschlechtsverteilung vorhandener Bewohner}
\usage{
pie.bewohner(Daten)
}
\arguments{
\item{Daten}{}
}
\description{
Piechart zu Geschlechtsverteilung vorhandener Bewohner
}
