% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter.bewohner.miete}
\alias{scatter.bewohner.miete}
\title{Scatterplot Miete nach Bewohnerzahl}
\usage{
scatter.bewohner.miete(Daten)
}
\arguments{
\item{Daten}{}
}
\description{
Scatterplot Miete nach Bewohnerzahl
}
