% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter.bewohner.mieteqm}
\alias{scatter.bewohner.mieteqm}
\title{Scatterplot Miete pro Quadratmeter nach Bewohnerzahl}
\usage{
scatter.bewohner.mieteqm(Daten)
}
\arguments{
\item{Daten}{}
}
\description{
Scatterplot Miete pro Quadratmeter nach Bewohnerzahl
}
