% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter.geschl.mieteqm}
\alias{scatter.geschl.mieteqm}
\title{Scatterplot Miete pro Quadratmeter nch Geschlecht}
\usage{
scatter.geschl.mieteqm(Daten)
}
\arguments{
\item{Daten}{}
}
\description{
Scatterplot Miete pro Quadratmeter nch Geschlecht
}
